Changes to allow installation on Ubuntu 18 and other minor releases from 14, 16.
Updates dvd.sh to install net-tools as it is currently assumed to be installed.''''

Added conditionals for ubuntu 18.

Updated `README.md` with instructions that will work on any of the major releases supported.

Updated `README.md` to add a masterkey configuration example as was unclear having only included field names.